Enhance dashboard
