+++ website/src/constants/path.ts
@@ -2,17 +2,17 @@ const BASE_PATH = "dashboard";
 
 export const PATH_DASHBOARD = {
   root: "/",
-  home: `/${BASE_PATH}/main-results`,
+  home: `/${BASE_PATH}`,
   benchmarkSet: {
-    list: `/${BASE_PATH}/benchmark-set`,
+    list: `/${BASE_PATH}#benchmark-set`,
     one: `/${BASE_PATH}/benchmark-set/{name}`,
   },
-  benchmarkSummary: `/${BASE_PATH}/benchmark-summary`,
+  benchmarkSummary: `/${BASE_PATH}#benchmark-summary`,
   featureDistribution: `/${BASE_PATH}/benchmark-set/feature-distribution`,
-  compareSolvers: `/${BASE_PATH}/compare-solvers`,
-  solvers: `/${BASE_PATH}/solvers`,
-  performanceHistory: `/${BASE_PATH}/performance-history`,
-  fullResults: `/${BASE_PATH}/full-results`,
+  compareSolvers: `/${BASE_PATH}#compare-solvers`,
+  solvers: `/${BASE_PATH}#solvers`,
+  performanceHistory: `/${BASE_PATH}#performance-history`,
+  fullResults: `/${BASE_PATH}#full-results`,
 };
 
 export const ROOT_PATH = {
